Modified port configuration, added application source file: measure.c

+++ Includes/adc.h
@@ -13,6 +13,6 @@
 void adcInit(void);
 void adcStart(void);
 char adcIsFinished(void);
-uint16_t adcMeasure(void);
+uint32_t adcMeasure(void);
 
 #endif /* ADC_H_ */
